Caret layer should focus on showing the caret, not include movng the caret.
